refactor : days 리팩터링
