+++ src/components/pages/wishlist/List.tsx
@@ -15,7 +15,7 @@ import Spinner from "@/components/public/Spinner/Spinner"
 import { location } from "@/constants/meeting"
 import useWishList from "@/hooks/useWishList"
 import { IFilterOption } from "@/types/meeting/meeting"
-import dayjs from "dayjs"
+import { isCurrentDateAfter } from "@/util/days"
 
 const List = () => {
   const [filter, setFilter] = useState<IFilterOption>({
@@ -118,7 +118,7 @@ const List = () => {
           {wish.map((list) => {
             return (
               <div key={list.id} className="relative mt-6 first:mt-0">
-                {dayjs().isAfter(dayjs(list.registrationEnd)) && (
+                {isCurrentDateAfter(list.registrationEnd) && (
                   <div className="absolute left-0 top-0 z-10 flex h-full w-full flex-col items-center justify-center gap-6 rounded-3xl bg-black/80 text-center text-sm font-medium leading-5 text-white sm:flex-row">
                     마감된 챌린지에요, <br />
                     다음 기회에 만나요 🙏
